+++ src/models/index.ts
@@ -1,6 +1,7 @@
 export type { default as AppSettings } from './settings';
 export type { default as Merchant, MerchantId } from './merchant';
 export type { default as Note, NoteId } from './note';
+export type { default as Notification, NotificationId } from './notification';
 export type { default as Post, PostId } from './post';
 export type { default as Product, ProductId } from './product';
 export type { default as Profile, ProfileId } from './profile';