+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Add `LocationQueryBottomSheet` component with basic UI
 - Display Parse server URL on the settings page for ease of troubleshooting
-- Add placeholder `NotificationsScreen`
+- Add basic support for notifications with a `NotificationsScreen` component
 
 ### Fixed